Fix manifest-symbol matching.
In the process resource task we need to pass pairs of
manifest file and symbol file (R.txt). Both these come from
2 different FileCollection using the new Gradle API.

I was trying to match them using ComponentArtifactIdentifier which
is wrong. ComponentIdentifier was what's needed. This is because
the former contains not only information about the component (the
project) but also the artifact (the manifest or the R.txt) so
it didn't match.

Test: none.
Change-Id: I8ebc60852b8653b259669631c482fafb14e2349f