Fix API check to handle workaround with interface
Activity#isDestroyed was added in API 17. Some apps work around this
by adding an interface which declares the isDestroyed() method, and
then have their activities all implement the interface.

Lint was not correctly handling this: when resolving methods, it just
resolved the actual called method for the current API level (e.g.
Activity#isDestroyed()) and then complained if the minSdkVersion is
not at least 17.

With this CL, it now looks up interface methods that also match the
method call, and if any of those have a lower API requirement, then
that one is used instead.

Test: New unit test included
Change-Id: Ibe514b9de00137173c34793565415786020689cd