Fix issue with test app with android dependencies.
A copy-paste issue marked the android transitive dependencies
of a tested app (via a separate module) as skipped instead of
provided on the compile config, removing them completely from
the graph.
Because of how the recursive method worked, the direct dependency
of the tested app would be fine. It's only deep ones that were
marked erronously as skipped.

Package dependencies are marked skipped while compile dependencies
must be marked provided.

Also fix how the transitive dependendencies of the tested app
show up in the full graph of the test app. We used to not
be able to have JARs with AAR children, so we'd pull them
up to the same level as the JAR. Now that we can (due
to the introduction of DependencyNode), the old code
was still there but broken. It both pulled up the android
dependencies but also left them as transitive dependencies
of the JAR too. I fixed this which required fixing a test
that tested the model incorrectly.

bug: https://code.google.com/p/android/issues/detail?id=232843
Test: Updated a new test to show the issue.
Change-Id: I48c32bc96c9c896de4712da1e179373cee2ad631