Start using the usage attributes.
This allows the compile and package(runtime) configuration
on the consuming side to consume a different graph. The producer
can tweak each graph to be different to reduce the content of the
api (compile) graph to hide some implementation dependencies.

For now this is not used when publishing libraries, but on the
consuming side we want to use it, in particular to consume
java-library's separate output for classes and resources
rather than the jar that contains both.

This also adds soem additional debug output to look at published
variant-specific artifacts.

Test: existing ones.
Change-Id: I81c840447cdb47cfb33970844ee4cf5606bee5a8