Use different enums for consumed and published configurations
Previously ConfigType attempted to reprensent both the configs
used for publishing and the ones used for consuming. While there
is a connection between api/compile and runtime/runtime, this
breaks with the annotation processor config.

It was also confusing to say "publish to compile" which is not
quite accurate. This separation makes things more clear and allows
to enforce that we only consume for the annotation processor something
published to runtime_elements.

Test: existing.
Change-Id: Ib0d6ed7466191848ac33bb57395f81fb38cd5e55