Clean out ouput before non-inc transform.
The transforms that copies classes.jar file and the local
jar dependencies of aar into the bundle folder didn't clean
out the local jar folder before running even though the transform
is not incremental.

A removal of a local dependencies, or a change in the default
variant (when not publishing all variant) could lead to some
remaining files in the local jar folders that should not be there
anymore.

bug: https://code.google.com/p/android/issues/detail?id=233777
Test: none, need to write a new one.
Change-Id: Ic85d894f3b6f1ffaf035e802641731caa0ce88d3