Remove obsolete tests.
AppWithCompileLocalAarFromOlderIdeTest
This test checks what happens when an older IDE (Studio 1.0)
opens the project. This will not be supported any longer
(will require > 2.4 AS) so this test is obsolete.

AppWithPackageLibTest
This test is about a runtimeOnly dependency on an AAR
which was not supported before. It's now fine.

AppWithPackageLocalAarTest
Local AARs are now supported. AARs in runtimeOnly dependencies
are supported too. Obsolete test.

Test: fixing (removing) them.
Change-Id: I01fdcf1a79df879ae4e2404d89e2d3e656282b40