Fix merged manifest source locations
Lint has APIs for letting detectors access the merged manifest, and for
computing the original source locations when errors are reported against
nodes in the merged manifest.

However, this was not working correctly in real world scenarios: when there
is just one manifest file (as is the case for most tests), it was just
passing through the source manifest as the merged manifest, and this manifest
happens to have the same path as the merged manifest. But in the general case
the merged manifest and the source manifest are generally not the same, and
we need to pass around the source file path too, not just the DOM node. This
required tweaks in a number of files.

This CL also makes manifest merging able to use the manifest merger's action
log directly, instead of just using json blame files.

Test: The unit test for merged manifest lookup was updated. This functionality
is also tested indirectly by an upcoming lint check which uses the merged
manifest more deeply (which was what exposed this limitation in the first place.)

Change-Id: If269a2fa7be3e629741952ea37a3f4c539d9884a