Adding AgentStatusChanged callback to perfd.
This change adds a callback listener to perfd for when the agent
status changes, and hooks up that callback to the events profiler.
When a perfa agent dies the status changed callback gets called.
EventService is currently the only consumer of this, and uses it to
force kill any activities that were left open.

b/35527754

Test: None

Change-Id: I5da9d85854689dceff7da810a1d1ad84b1f84ac3