Force Lint to run in headless mode
IconDetector uses ImageIO in a way which triggers the initialization
of java.awt.GraphicsEnvironment in the classloader set by the Gradle
daemon, and that code checks again for the headless property.

It looks like the system properties are reset at some point (we do
pass -Djava.awt.headless=true to the Gradle daemon, and it stays like
that during sync, etc). But running the lint task shows the Java icon
in the dock on OS X.

Bug: http://b.android.com/234791
Test: manual for now
Change-Id: I4eea9cc796f03cc72663e0ed3d0105cd7c4a4b85