Use API level of target device only in specific cases.
GradleVariantConfiguration.getMinSdkVersion() currently overrides
VariantConfiguration.getMinSdkVersion(). If the API level of the target
device is passed from the IDE, that method computes minSdkVersion based
on that injected value, which can cause runtime errors as in
http://b/37133933 or http://b/37140083.

We should not let GradleVariantConfiguration.getMinSdkVersion() override
VariantConfiguration.getMinSdkVersion() and should use it only in
multi-dex mode and in specific cases.

We may need to come up with a comprehensive solution regarding
minSdkVersion at some point.

Bug: http://b/37133933, http://b/37140083
Test: GradleVariantConfigurationTest, DeploymentApiOverrideTest,
    MergeResourcesTest#injectedMinSdk
Change-Id: I3ade4e51f4ebb8bccaaa1803dc50c42272cf9a3a