Fix IconDetector path handling
The IconDetector was using File.separator to display icon name
and folder configuration combinations, e.g. on Windows it
attempted to show "drawable-mdpi\next.png". However, the
backslash was passed to String.format which then would interpret the \
as an escape - \n in this example.

This fix pulls the filename formatting outside the String.format usage
intended for number formatting only.

Test: Already covered by tests (which were failing on Windows)

Change-Id: Ib14bddaf10bd38077b8ba70e56b4916f4776e75f