Use Unix abstract sockets with JVMTI-enabled agents.
Unix abstract socket doesn't work between a shell user (perfd)
and an app user (agent). Therefore, perfd will first create a
client socket connecting to itself, then use execl() and run-as
to run a new process ('perfd --connect'). The new connector
process inherits the client socket's fd (thanks to excl()) and
is able to pass it to agent (thanks to run-as).

In follow-up CLs, we need to handle cases like:
* perfd restarts but the app remains running
* multiple app start at the same time, each should have a
  dedicated connector.

Bug: 35672396

Test: manual
<enable.jvmti.profiling=true/false> X <O and older devices>

Change-Id: I51d72ec1e87f2b229ef286c20717ee02481ccaf6