Fix API check for the support library
In the past, with the bytecode based API check implementation, lint
had to check virtual dispatch on its own.

For example, if you have this code:

    void myMethod(AppCompatActivity activity) {
       activity.method();
    }

the bytecode would say that this is calling method "method" on class
"AppCompatActivity", but it may be an inherited method, so lint would
need to walk up the inheritance hierarchy for and check each super
class until it found a match.

This is no longer necessary (because in the PSI based world we resolve
the method call at compilation time), and the old machinery was
causing false positives when for example checking the support library
source code itself.

This CL removes not just the support-library workarounds, but the
whole bytecode based API check machinery too since it's no longer
used.

Test: Unit tests updated

Change-Id: I557d99872bba6a18e2e41d742d849e7993e209e0