Fix lint unit tests on Windows
An earlier commit disabled all the failing lint tests on Windows.
This restores these tests and fixes the problems. The main change is
to create a new utility method for formatting relative paths (parent
folder + file name) which is used in a number of places, and this
utility method then is subclassed in tests to do stable Unix file
separators. I then removed all the hacks where all the output
converted backslashes to forward slashes, which caused a lot of
breakage (and bad looking output when the backslashes were unrelated
to paths, for example displaying warnings about string escapes, \n.)

Test: Yes, that's what this fixes

Change-Id: I067951de1e5269deede4735222bd31fa021fbf43